requirements.txt
# ASV-INSPECT Requirements
# Automated Statistical Verification for Assembly Inspection

# Core Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Deep Learning
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Handling
pyyaml>=6.0
pydantic>=2.0.0

# Optional but recommended
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
